Resolve dir path in early stage
Linkerconfig generator sorts dir path in revert order, so when a dir is
subdirectory of other dir, then subdirectory can be mentioned earlier.
However, this sorting occured before resolving environmental variables,
so this list was failing to be kept in actual reverse order. To fix this
issue, dir path should be resolved before it is collected by generator.

Test: m -j && tested from cuttlefish
Change-Id: I21f88cca25fef37e080dc1c1cb68e7a0c0186b4f