Create full test for ld.config.txt content
There are some module level unit tests and backward compatibility tests,
but still there is no test that verifies ld.config.txt content is
valid in terms of syntax and also context. To test both
features in basic level, this change adds full content test which parses
content and checks if every context (such as path exists and links
between namespace is valid) is able to be used by linker.

Bug: 139639353
Test: m -j && atest passed
Change-Id: I58255ccc6602f6bcb86b897fc89f5581ccc18f08