Add runtime namespace
libc_malloc_hooks/debug.so are included in the runtime APEX, which now
is only for the bionic libs. The libs should be loaded in isolatation
from other libs because they might not be compatible with other libs in
/system/lib. (the libs depend on libc++.so)

Creating the namespace 'runtime' where the libs can be loaded into.

No modification to the bionic side is needed because it already has been
trying to load the libs from the namespace. The namespace actually was
accidentaly disappeared during the split of the runtime APEX into
runtime and art APEXes.

Bug: 143892776
Test: bionic-unit-tests
Change-Id: I638c5f0439eb402338b90f2c01fddbfdae7d73e9