Define vndk_product namespace for product apps
Unbundled apps may use vndk-sp libs from the system section. However,
product apps must use the vndk-sp libs for the product vndk version
that can be different from the vendor vndk version. So we must define
a vndk namespace for product partition.

For this purpose, BuildVndkNamespace() function requires an
additional parameter to indicate which partition is using the
namespace.
Using this function, the system section has an additional namespace
called "vndk_product" which is used only for unbundled product apps.

Test: atest linkerconfig_modules_unittest
Bug: 149063221
Change-Id: I3658efdc75e5c9a0ac3b92b9afc68aff2d4cb300