Generate ASAN paths with certain rules
Previously each paths had to specify if it should be included in ASAN
paths, and /data/asan is valid. However, most of paths comes with ASAN
path with /data/asan except /apex, we can make this as simple so anyone
can add extra paths without understanding of ASAN. This change is the
first step to remove all ASAN path description from path definition.

Bug: 139788162
Test: m -j passed && Crosshatch HWASAN boot succeeded
Change-Id: I7e89cfd81e5b5699112b0636691a8b6007a98f33