Create isolated section under /data
Some of the engineers are using /data for non-framework development, but
it is hard to use as /data shares section with /system, so many
libraries are linked with system one first. This changes creates a new
isolated section from subdirectory of /data so developers can run their
own executables with linked to their own libs first.

Bug: 160231064
Test: Checked from Cuttlefish that isolated section is created as expected
Change-Id: I047ee69b9ae7a4b01894c3a29efd4226ab47ce56