Drop AllowAllSharedLibs from the ART APEX namespace.
It is an isolation problem, and due to an overcaching issue in the
linker (b/142941232) it can also result in system libraries overriding
libraries with the same name in the ART APEX when they are loaded from
the ART linker namespace.

With fixes in ART to use the system namespace rather than its internal
APEX namespace, this can now be dropped.

Test: system/linkerconfig/rundiff.sh
Bug: 130340935
Change-Id: I334846e72f1ad9366e933795dc4763e990b9d96c